requirements.txt
click>=6.7,<7.0
PyYAML>=3.13,<4.0